Reverted partial 8c8b933eaff4c22d9258c30b1b86a1a18ede2667
